requirements.txt
discord.py
requests
python-dotenv
sqlalchemy>=2.0.0
pytest
pytest-asyncio
pytest-xdist
google-cloud-secret-manager
cloud-sql-python-connector[pg8000]
aiohttp
colorama
